# docs/swagger.yaml
basePath: /api/
definitions:
...
  background.Background:
    properties:
      abilities:
        items:
          type: string
        type: array
      name:
        type: string
      requirements:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: meshee.knight@gmail.com
    name: Meshenka
  description: Instant random character generator
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Who my f*cking Nimble 5e character is?
  version: "1.0"
paths:
...
  /backgrounds/{name}:
    get:
      description: Get a background by name
      parameters:
      - description: Background name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/background.Background'
        "404":
          description: Not Found
      summary: One Background
      tags:
      - background
...
swagger: "2.0"